bugfix in quantification
git-svn-id: http://svn.code.sf.net/p/qlandkartegt/code/QLandkarteGT/trunk@1317 d0c010d8-bb82-477e-a075-112140230451

+++ src/CPlotAxis.h
@@ -81,7 +81,7 @@ class CPlotAxis : public QObject
             if ( scale == 0 ) {
                 return 0;
             }
-            return ( double ) ( ( pt -0.5 ) / scale + used_min );
+            return ( double ) ( ( (double)pt - 0.5 ) / scale + used_min );
         }
 
         void setAutoscale(bool on){autoscale = on;}
